ci(gql): rm gql schema file
The need for keeping the GQL schema as a file in the repo has been removed by the addition of `scripts/codegen.types-graphql.sh`. GQL-schema types are now generated from the actual source of truth - the remote schema - rather than a facsimile which can be out of date. Rather than updating the schema file every time when the associated codegen is run, fresh schema info is instead just pulled every time from the remote source of truth.