fix: add isAuthInitComplete state
This additional state ensures the hook never runs more than once. During dev, the hook would re-run if the user was initially auth'd and then logged out, which on Chrome was causing the Google OAuth window to then be shown and subsequently logging them back in if the Google OAuth flow succeeded. This state ensures there's no logout-->login loop.